@@ -8,50 +8,56 @@
     "url": "https://github.com/seaofvoices/darklua/issues"
   },
   "dependencies": {
-    "darklua": "./darklua-wasm/pkg",
-    "@babel/eslint-parser": "^7.22.15",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.14.14",
-    "acorn": "^8.10.0",
-    "acorn-import-assertions": "^1.8.0",
+    "@babel/eslint-parser": "^7.23.10",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
+    "@fontsource-variable/comfortaa": "^5.0.18",
+    "@mui/icons-material": "^5.15.13",
+    "@mui/material": "^5.15.13",
+    "acorn": "^8.11.3",
+    "acorn-import-assertions": "^1.9.0",
     "babel-plugin-prismjs": "^2.1.0",
-    "gatsby": "^5.12.7",
-    "gatsby-plugin-catch-links": "^5.12.0",
-    "gatsby-plugin-image": "^3.12.1",
-    "gatsby-plugin-manifest": "^5.12.1",
-    "gatsby-plugin-material-ui": "^4.1.0",
-    "gatsby-plugin-react-helmet": "^6.12.0",
-    "gatsby-plugin-sharp": "^5.12.1",
-    "gatsby-remark-copy-linked-files": "^6.12.0",
-    "gatsby-remark-images": "^7.12.1",
-    "gatsby-remark-prismjs": "^7.12.0",
-    "gatsby-remark-responsive-iframe": "^6.12.0",
-    "gatsby-remark-smartypants": "^6.12.0",
-    "gatsby-source-filesystem": "^5.12.0",
-    "gatsby-theme-material-ui": "^5.3.0",
-    "gatsby-transformer-remark": "^6.12.1",
-    "gatsby-transformer-sharp": "^5.12.1",
-    "joi": "^17.11.0",
+    "darklua": "./darklua-wasm/pkg",
+    "gatsby": "^5.13.3",
+    "gatsby-plugin-catch-links": "^5.13.1",
+    "gatsby-plugin-image": "^3.13.1",
+    "gatsby-plugin-manifest": "^5.13.1",
+    "gatsby-plugin-react-helmet": "^6.13.1",
+    "gatsby-plugin-sharp": "^5.13.1",
+    "gatsby-remark-copy-linked-files": "^6.13.1",
+    "gatsby-remark-images": "^7.13.1",
+    "gatsby-remark-prismjs": "^7.13.1",
+    "gatsby-remark-responsive-iframe": "^6.13.1",
+    "gatsby-remark-smartypants": "^6.13.1",
+    "gatsby-source-filesystem": "^5.13.1",
+    "gatsby-transformer-remark": "^6.13.1",
+    "gatsby-transformer-sharp": "^5.13.1",
+    "joi": "^17.12.2",
     "json5": "^2.2.3",
-    "mdast-util-from-markdown": "^1.3.0",
-    "mdast-util-to-hast": "^12.3.0",
-    "monaco-editor": "^0.44.0",
+    "mdast-util-from-markdown": "^2.0.0",
+    "mdast-util-to-hast": "^13.1.0",
+    "monaco-editor": "^0.47.0",
     "monaco-editor-webpack-plugin": "^7.1.0",
     "prismjs": "^1.29.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-use-ref-effect": "^1.2.0",
-    "rehype-react": "^7.2.0",
-    "typeface-merriweather": "1.1.13",
-    "typeface-montserrat": "1.1.13",
-    "typescript": "^4.9.5"
+    "rehype-react": "^8.0.0",
+    "typescript": "^5.4.2",
+    "unified": "^11.0.4"
   },
   "devDependencies": {
-    "babel-preset-gatsby": "^3.9.0",
-    "prettier": "^2.8.8"
+    "babel-preset-gatsby": "^3.13.1",
+    "prettier": "^3.2.5"
+  },
+  "optionalDependencies": {
+    "@parcel/watcher-linux-x64-glibc": "^2.4.1"
+  },
+  "overrides": {
+    "react-server-dom-webpack@0.0.0-experimental-c8b778b7f-20220825": {
+      "react": "^18.2.0"
+    }
   },
   "homepage": "https://github.com/seaofvoices/darklua",
   "keywords": [
