@@ -11,8 +11,8 @@ use crate::{nodes::Identifier, process::NodeProcessor};
 /// # use darklua_core::nodes::Expression;
 /// # use darklua_core::process::processors::FindVariables;
 /// # use darklua_core::process::{DefaultVisitor, NodeProcessor, NodeVisitor};
-/// let variables = vec!["foo".to_owned()];
-/// let mut find_foo: FindVariables = variables.iter().collect();
+/// let variables = vec!["foo"];
+/// let mut find_foo: FindVariables = variables.into_iter().collect();
 ///
 /// let mut foo_expression = Expression::identifier("foo");
 /// DefaultVisitor::visit_expression(&mut foo_expression, &mut find_foo);
@@ -25,27 +25,34 @@ use crate::{nodes::Identifier, process::NodeProcessor};
 /// # use darklua_core::nodes::Expression;
 /// # use darklua_core::process::processors::FindVariables;
 /// # use darklua_core::process::{DefaultVisitor, NodeProcessor, NodeVisitor};
-/// # let variables = vec!["foo".to_owned()];
-/// # let mut find_foo: FindVariables = variables.iter().collect();
+/// # let variables = vec!["foo"];
+/// # let mut find_foo: FindVariables = variables.into_iter().collect();
 /// let mut bar_expression = Expression::identifier("bar");
 /// DefaultVisitor::visit_expression(&mut bar_expression, &mut find_foo);
 ///
 /// assert!(!find_foo.has_found_usage());
 /// ```
 pub struct FindVariables<'a> {
-    variables: Vec<&'a String>,
+    variables: Vec<&'a str>,
     usage_found: bool,
 }
 
 impl<'a> FindVariables<'a> {
+    pub fn new(variable: &'a str) -> Self {
+        Self {
+            variables: vec![variable],
+            usage_found: false,
+        }
+    }
+
     #[inline]
     pub fn has_found_usage(&self) -> bool {
         self.usage_found
     }
 }
 
-impl<'a> FromIterator<&'a String> for FindVariables<'a> {
-    fn from_iter<T: IntoIterator<Item = &'a String>>(iter: T) -> Self {
+impl<'a> FromIterator<&'a str> for FindVariables<'a> {
+    fn from_iter<T: IntoIterator<Item = &'a str>>(iter: T) -> Self {
         Self {
             variables: iter.into_iter().collect(),
             usage_found: false,
