@@ -18,6 +18,8 @@ pub trait Scope {
     /// Called when entering a function block (with each parameters of the function), with the
     /// identifiers from a generic for statement or the identifier from a numeric for loop.
     fn insert(&mut self, identifier: &mut String);
+    /// Called when entering a function defined with a method
+    fn insert_self(&mut self);
     /// Called when a new local variable is initialized.
     fn insert_local(&mut self, identifier: &mut String, value: Option<&mut Expression>);
     /// Called when a new local function is initialized.
@@ -36,13 +38,7 @@ impl ScopeVisitor {
             .for_each(|statement| Self::visit_statement(statement, scope));
 
         if let Some(last_statement) = block.mutate_last_statement() {
-            scope.process_last_statement(last_statement);
-
-            if let LastStatement::Return(expressions) = last_statement {
-                expressions
-                    .iter_mut_expressions()
-                    .for_each(|expression| Self::visit_expression(expression, scope));
-            };
+            Self::visit_last_statement(last_statement, scope);
         };
     }
 }
@@ -61,6 +57,13 @@ impl<T: NodeProcessor + Scope> NodeVisitor<T> for ScopeVisitor {
             .iter_mut_values()
             .for_each(|value| Self::visit_expression(value, scope));
 
+        for r#type in statement
+            .iter_mut_variables()
+            .filter_map(TypedIdentifier::mutate_type)
+        {
+            Self::visit_type(r#type, scope);
+        }
+
         statement.for_each_assignment(|variable, expression| {
             scope.insert_local(variable.mutate_name(), expression)
         });
@@ -68,6 +71,21 @@ impl<T: NodeProcessor + Scope> NodeVisitor<T> for ScopeVisitor {
 
     fn visit_function_expression(function: &mut FunctionExpression, scope: &mut T) {
         scope.process_function_expression(function);
+
+        for r#type in function
+            .iter_mut_parameters()
+            .filter_map(TypedIdentifier::mutate_type)
+        {
+            Self::visit_type(r#type, scope);
+        }
+
+        if let Some(variadic_type) = function.mutate_variadic_type() {
+            Self::visit_function_variadic_type(variadic_type, scope);
+        }
+
+        if let Some(return_type) = function.mutate_return_type() {
+            Self::visit_function_return_type(return_type, scope);
+        }
 
         scope.push();
         function
@@ -83,7 +101,25 @@ impl<T: NodeProcessor + Scope> NodeVisitor<T> for ScopeVisitor {
         scope.process_function_statement(statement);
         scope.process_variable_expression(statement.mutate_function_name().mutate_identifier());
 
+        for r#type in statement
+            .iter_mut_parameters()
+            .filter_map(TypedIdentifier::mutate_type)
+        {
+            Self::visit_type(r#type, scope);
+        }
+
+        if let Some(variadic_type) = statement.mutate_variadic_type() {
+            Self::visit_function_variadic_type(variadic_type, scope);
+        }
+
+        if let Some(return_type) = statement.mutate_return_type() {
+            Self::visit_function_return_type(return_type, scope);
+        }
+
         scope.push();
+        if statement.get_name().has_method() {
+            scope.insert_self();
+        }
         statement
             .mutate_parameters()
             .iter_mut()
@@ -97,6 +133,21 @@ impl<T: NodeProcessor + Scope> NodeVisitor<T> for ScopeVisitor {
         scope.process_local_function_statement(statement);
 
         scope.insert_local_function(statement);
+
+        for r#type in statement
+            .iter_mut_parameters()
+            .filter_map(TypedIdentifier::mutate_type)
+        {
+            Self::visit_type(r#type, scope);
+        }
+
+        if let Some(variadic_type) = statement.mutate_variadic_type() {
+            Self::visit_function_variadic_type(variadic_type, scope);
+        }
+
+        if let Some(return_type) = statement.mutate_return_type() {
+            Self::visit_function_return_type(return_type, scope);
+        }
 
         scope.push();
         statement
@@ -119,6 +170,13 @@ impl<T: NodeProcessor + Scope> NodeVisitor<T> for ScopeVisitor {
             .iter_mut_identifiers()
             .for_each(|identifier| scope.insert(identifier.mutate_name()));
 
+        for r#type in statement
+            .iter_mut_identifiers()
+            .filter_map(TypedIdentifier::mutate_type)
+        {
+            Self::visit_type(r#type, scope);
+        }
+
         Self::visit_block(statement.mutate_block(), scope);
     }
 
@@ -131,6 +189,10 @@ impl<T: NodeProcessor + Scope> NodeVisitor<T> for ScopeVisitor {
         if let Some(step) = statement.mutate_step() {
             Self::visit_expression(step, scope);
         };
+
+        if let Some(r#type) = statement.mutate_identifier().mutate_type() {
+            Self::visit_type(r#type, scope);
+        }
 
         scope.push();
         scope.insert(statement.mutate_identifier().mutate_name());
@@ -220,6 +282,10 @@ impl Scope for IdentifierTracker {
         self.insert_identifier(identifier);
     }
 
+    fn insert_self(&mut self) {
+        self.insert_identifier("self");
+    }
+
     fn insert_local(&mut self, identifier: &mut String, _value: Option<&mut Expression>) {
         self.insert_identifier(identifier);
     }
@@ -248,6 +314,11 @@ where
     #[inline]
     fn insert(&mut self, identifier: &mut String) {
         self.deref_mut().insert(identifier);
+    }
+
+    #[inline]
+    fn insert_self(&mut self) {
+        self.deref_mut().insert_self();
     }
 
     #[inline]
