@@ -7,7 +7,7 @@ use crate::rules::{
 use super::verify_no_rule_properties;
 
 #[derive(Debug, Default)]
-struct RemoveCommentProcessor {}
+pub(crate) struct RemoveCommentProcessor {}
 
 impl NodeProcessor for RemoveCommentProcessor {
     fn process_block(&mut self, block: &mut Block) {
